requirements.txt
Django==1.3
psycopg2==2.4.5
dj-database-url==0.2.0
BeautifulSoup
python-dateutil
Markdown
django-improved-inlines
django-tagging
django-registration
django-basic-apps
django-profiles
